modularized read, uppercase, and write
