+++ test.txt
@@ -1,1 +1,1 @@
-test string
+TEST STRING
